Transform Split - building lib-tempo
