+++ src/lib-tempo/CMakeLists.txt
@@ -9,6 +9,7 @@ set(TEMPO_LIB_TEMPO_HEADERS
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentGridAi.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentHealth.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentRender.hpp"
+	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/ComponentTransform.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/EntityCreation.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/EntityCreationClient.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/EntityCreationServer.hpp"
@@ -20,7 +21,6 @@ set(TEMPO_LIB_TEMPO_HEADERS
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/PlayerRemote.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/PlayerRemoteServer.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/TileMask.hpp"
-	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/Transform.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/network/base.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/network/client.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/network/queue.hpp"
@@ -28,6 +28,7 @@ set(TEMPO_LIB_TEMPO_HEADERS
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemGridAi.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemHealth.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemRender.hpp"
+	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/SystemTransform.hpp"
 )
 
 set(TEMPO_LIB_TEMPO_SOURCE
@@ -38,6 +39,7 @@ set(TEMPO_LIB_TEMPO_SOURCE
 	"${CMAKE_CURRENT_LIST_DIR}/src/time.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentHealth.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentRender.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/src/entity/ComponentTransform.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/EntityCreation.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/EntityCreationClient.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/EntityCreationServer.cpp"
@@ -49,7 +51,6 @@ set(TEMPO_LIB_TEMPO_SOURCE
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/PlayerRemote.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/PlayerRemoteServer.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/TileMask.cpp"
-	"${CMAKE_CURRENT_LIST_DIR}/src/entity/Transform.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/network/base.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/network/client.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/network/queue.cpp"
@@ -57,6 +58,7 @@ set(TEMPO_LIB_TEMPO_SOURCE
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemGridAi.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemHealth.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemRender.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/src/entity/SystemTransform.cpp"
 )
 
 add_library(lib_tempo STATIC
