+++ src/lib-tempo/src/entity/SystemTransform.cpp
@@ -1,21 +1,7 @@
-#include <tempo/entity/Transform.hpp>
+#include <tempo/entity/SystemTransform.hpp>
+#include <tempo/entity/ComponentTransform.hpp>
 
 namespace tempo{
-	ComponentTransform::ComponentTransform() : ComponentTransform(0, 0, 0){
-		// empty body
-	}
-
-	ComponentTransform::ComponentTransform(Ogre::Vector3 pos) : ComponentTransform(pos.x, pos.y, pos.z){
-		// empty body
-	}
-
-	ComponentTransform::ComponentTransform(Ogre::Real x, Ogre::Real y, Ogre::Real z) {
-		this->position.x = x;
-		this->position.y = y;
-		this->position.z = z;
-		this->rotation= Ogre::Quaternion();
-	}
-
 
 	void SystemUpdateTransforms::update(SystemLevelManager& level){
 		auto entities = getEntities();
